feat: 반응형 헤더 및 Layout 구조 적용 및 헤더 리펙토링
