+++ apps/farminglog/package.json
@@ -12,7 +12,9 @@
   },
 "dependencies": {
     "@repo/api": "workspace:*",
-    "@repo/auth": "workspace:*"
+    "@repo/auth": "workspace:*",
+    "@repo/router": "workspace:*"
+
   },
   "devDependencies": {}
 }
