+++ apps/farminglog/src/pages/home/index.tsx
@@ -1,7 +1,6 @@
 import Harvest from "./Harvest/harvest";
 import Ranking from "./Ranking/ranking";
 import Cheer from "./Cheer/cheer";
-import Header from "../Component/Header";
 import * as S from "./index.styled";
 import useMediaQueries from "../../../../website/src/hooks/useMediaQueries";
 
@@ -9,7 +8,6 @@ export default function Main() {
   const { isApp, isMobile } = useMediaQueries();
   return (
     <>
-      <Header/>
       <S.MainContainer $isApp={isApp} $isMobile={isMobile}>
         <Harvest />
         <Ranking />
