+++ apps/farminglog/src/router/AppRouter.tsx
@@ -1,4 +1,3 @@
-// AppRouter.tsx
 import { RouterProvider } from "react-router";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { router } from "./index";
